Bryan: Click Here For Description
Added a long enemy cube called Timed Cube Enemy behind the player sphere.

Added a script for the TImed Cube Enemy; it moves right.

Added more code to the sphere's script, now when sphere touches the Timed Cube Enemy, it resets the sphere to the starting position and the timed cube as well.

Changed the scene name to Bryan Level.